+++ src/main/java/com/dreamcloud/esa_score/fs/FileSystem.java
@@ -2,7 +2,7 @@ package com.dreamcloud.esa_score.fs;
 
 public class FileSystem {
     public static int TERM_LENGTH_BYTES = Integer.BYTES;
-    public static int OFFSET_BYTES = Integer.BYTES;
+    public static int OFFSET_BYTES = Long.BYTES;
     public static int SCORE_BYTES = Float.BYTES;
     public static int DOCUMENT_SCORE_BYTES = Integer.BYTES + SCORE_BYTES;
 }
