+++ src/main/java/com/dreamcloud/esa_score/fs/TermIndexEntry.java
@@ -3,6 +3,6 @@ package com.dreamcloud.esa_score.fs;
 public class TermIndexEntry {
     public String term;
     public int documentFrequency;
-    public int offset;
+    public long offset;
     public int numScores;
 }
